Changed default DRB id from 3 to 1 With folowwing implicit rule drb id = x => correspondind lc id = x+2, eps bearer id = x+4
Changed RLC inner structs, no more statically allocated array, now it is dynamic via malloc and collection of RLC entities are stored in a hashtable.
Changed misc simple types
Started compiling with all warning setted and treated as errors, still working on RLC and around...
Add const qualifier wherever suitable (for example param frame, module id, const pointers, etc)

Thinking about a type of context parameter in procedure calls in L2 data plane (modules id, frame, srb flag, enb flag, etc)...


git-svn-id: http://svn.eurecom.fr/openair4G/trunk@5222 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair2/LAYER2/PDCP_v10.1.0/pdcp_netlink.c
@@ -204,7 +204,7 @@ void *pdcp_netlink_thread_fct(void *arg) {
            * NOTE: netlink messages can be assembled to form a multipart message
            */
           for (nas_nlh_rx = (struct nlmsghdr *) nl_rx_buf;
-              NLMSG_OK(nas_nlh_rx, len);
+              NLMSG_OK(nas_nlh_rx, (unsigned int)len);
               nas_nlh_rx = NLMSG_NEXT (nas_nlh_rx, len)) {
 
               /* There is no need to check for nlmsg_type because
